useEnter hook for list-item
